+++ src/main/java/org/apache/componentworld/entities/auto/_TapestryVersion.java
@@ -30,11 +30,11 @@ public abstract class _TapestryVersion extends CayenneDataObject {
         return (String)readProperty("description");
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         writeProperty("id", id);
     }
-    public Long getId() {
-        return (Long)readProperty("id");
+    public Integer getId() {
+        return (Integer)readProperty("id");
     }
 
     public void setName(String name) {
